Manca una piccolezza nel resize della finestra.

+++ trunk/elanja-fltk/regionCountStats.h
@@ -11,7 +11,7 @@ class regionCountStats : public glStats
 {
 public:
 	regionCountStats(int w,int h,const char *l=0);
-     	double regionHistory[STATS2_W];		
+     	int regionHistory[STATS2_W];		
 
 protected:
 	void paint();
